Code Splited : MangaElementDef2
