+++ src/mangadex/resources/componnents/mangas/v1/MangaElementDef2/error.tsx
@@ -18,7 +18,7 @@ export function ErrorBoundaryComp({error} : {
                             {error.name}
                         </AlertTitle>
                         <AlertDescription noOfLines={1}>
-                            {error.name}
+                            {error.message}
                         </AlertDescription>
                     </Alert>
                 </CardBody>
