+++ src/mangadex/resources/componnents/mangas/v1/MangaElementDef2/vanilla/Description.tsx
@@ -1,0 +1,46 @@
+import * as Chakra from "@chakra-ui/react";
+import { get_manga_description } from "@mangadex/resources/hooks/MangaStateHooks";
+import { useProps } from "../../MangaElementDef/vanilla";
+import ErrorEL1 from "@mangadex/resources/componnents/error/ErrorEL1";
+import React from "react";
+
+export default function Description() {
+    const { src } = useProps();
+    const {
+        manga_description_query
+    } = get_manga_description({
+        src
+    });
+    if (manga_description_query.isSuccess) {
+        if (manga_description_query.data.length == 0) {
+            return (
+                <React.Fragment />
+            );
+        } else {
+            return (
+                <Chakra.Text
+                    noOfLines={3}
+                    marginBottom={"1px"}
+                    fontSize={"md"}
+                >
+                    {manga_description_query.data[0].get_data()}
+                </Chakra.Text>
+            );
+        }
+    } else if (manga_description_query.isLoading) {
+        return (
+            <Chakra.Skeleton
+                height={"full"}
+                width={"full"}
+            />
+        );
+    } else if (manga_description_query.isError) {
+        return (
+            <ErrorEL1 error={manga_description_query.error} />
+        );
+    } else {
+        return (
+            <React.Fragment />
+        );
+    }
+}
