+++ src/mangadex/resources/componnents/mangas/v1/MangaElementDef2_withID.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { get_manga_byId, useMangaDownload_Delete } from "../../../hooks/MangaStateHooks";
+import { get_manga_byId } from "@mangadex/resources/hooks/MangaStateHooks";
 import ErrorEL1 from "../../error/ErrorEL1";
 import MangaFallback2 from "./MangaElement2Fallback";
 const MangaElementDef2 = React.lazy(() => import("./MangaElementDef2"));
@@ -8,28 +8,27 @@ export default function MangaElementDef2_withID(props: {
     mangaID: string
 }) {
     const { query } = get_manga_byId({
-        mangaID : props.mangaID
+        mangaID: props.mangaID
     });
-    if (query.isLoading) {
+    if (query.isSuccess) {
         return (
-            <MangaFallback2 />
+            <React.Suspense fallback={
+                <MangaFallback2 />
+            }>
+                <MangaElementDef2
+                    src={query.data.manga}
+                    isRefetching={query.isRefetching}
+                    refetch={query.refetch}
+                />
+            </React.Suspense>
         );
-    }
-    if (query.isError) {
+    } else if (query.isError) {
         return (
             <ErrorEL1 error={query.error} />
         );
-    }
-    return (
-        <React.Suspense fallback={
+    } else {
+        return (
             <MangaFallback2 />
-        }>
-            <MangaElementDef2
-                src={query.data!.manga}
-                isRefetching={query.isRefetching}
-                refetch={query.refetch}
-            />
-        </React.Suspense>
-
-    );
+        );
+    }
 }
